Make `guppi_fread_header` be DIRECTIO=1 aware
Inputs files that were written using Direct I/O will have headers that
have been padded (with bytes of arbitrary/unspecified values) to the
next multiple of 512 bytes.  They will also include a DIRECTIO=1 header
to indicate this.  The `guppi_fread_header` function is now aware of
this header variable and its significance and will skip past the padding
bytes after reading and storing the header if this header is present and
its value is 1.  If the DIRECTIO header is not present or its value is 0
(or anything that doesn't contain a 1) then `guppi_fread_header` will
know that there is no padding to skip.  In either case, the first byte
read after `guppi_fread_header` returns will be the first byte of the
data block that follows the header.