Makefile cleanup
Make the Makefile be more explicit about targets and dependencies and
which ones are real vs "phony" targets.  This means `make` will just
say:

    make: Nothing to be done for `all'.

when run after a successful completion of an earlier `make` command.

Also change the default to install into the `bin` and `lib` directories
that are adjacent to the `src` directory (rather than `/bin` and
`/lib`!).